requirements.txt
flask
python-dotenv
flask-restful
replicate
pyodbc
pandas
sqlalchemy
tabulate
Pylance
flask_sqlalchemy
flask-cors
pathlib